--- src/modules/recipes/types.ts
@@ -1,7 +1,0 @@
-export interface Recipe {
-  idMeal: string;
-  strMeal: string;
-  strMealThumb: string;
-  strInstructions: string;
-  [key: string]: any; // For other dynamic properties like ingredients
-}
